+++ KeeAnywhere/Configuration/ConfigurationServiceLocalSecureStore.cs
@@ -13,7 +13,9 @@ namespace KeeAnywhere.Configuration
 
         private void LoadAccountsFromLocalSecureStore()
         {
-            var path = AppConfigSerializer.LocalAppDataDirectory;
+            this.MigrateDataToRoamingAppData();
+
+            var path = ConfigurationInfo.SettingsDirectory;
             var filename = Path.Combine(path, ConfigurationFile_Accounts);
 
             if (!File.Exists(filename))
@@ -43,7 +45,7 @@ namespace KeeAnywhere.Configuration
 
         private void SaveAccountsToLocalSecureStore()
         {
-            var path = AppConfigSerializer.LocalAppDataDirectory;
+            var path = ConfigurationInfo.SettingsDirectory;
             var filename = Path.Combine(path, "KeeAnywhere.Accounts.json");
 
             var settings = new JsonSerializerSettings
@@ -56,5 +58,23 @@ namespace KeeAnywhere.Configuration
 
             File.WriteAllText(filename, configString);
         }
+
+        // v1.5.0: #113 Adjust account storage location
+        public void MigrateDataToRoamingAppData()
+        {
+            if (ConfigurationInfo.IsPortable) { return; }
+
+            var filenameOld = Path.Combine(AppConfigSerializer.LocalAppDataDirectory, ConfigurationFile_Accounts);
+            var filenameNew = Path.Combine(ConfigurationInfo.SettingsDirectory, ConfigurationFile_Accounts);
+
+            if (!File.Exists(filenameOld)) { return; }
+
+            try
+            {
+                File.Move(filenameOld, filenameNew);
+            }
+            catch { }
+        }
+
     }
 }